+++ HISTORY.rst
@@ -2,7 +2,14 @@
 History
 =======
 
+0.2.0 (2017-09-20)
+------------------
+
+* Possible to disable session timeout.
+* Dashboard panel fullscreen mode.
+
 0.1.0 (2017-09-12)
 ------------------
 
-* First release on PyPI.
+* First release
+